chore: fix env handling, crud and cross resource policies

+++ packages/amplify-cli/src/extensions/amplify-helpers/update-amplify-meta.ts
@@ -42,33 +42,41 @@ function moveBackendResourcesToCurrentCloudBackend(resources) {
   const tagCloudFilePath = pathManager.getCurrentTagFilePath();
 
   for (let i = 0; i < resources.length; i += 1) {
-    const sourceDir = path.normalize(path.join(pathManager.getBackendDirPath(), resources[i].category, resources[i].resourceName));
-    const targetDir = path.normalize(
-      path.join(pathManager.getCurrentCloudBackendDirPath(), resources[i].category, resources[i].resourceName),
-    );
+    if (resources[i].serviceType !== 'imported') {
+      const sourceDir = path.normalize(path.join(pathManager.getBackendDirPath(), resources[i].category, resources[i].resourceName));
+      const targetDir = path.normalize(
+        path.join(pathManager.getCurrentCloudBackendDirPath(), resources[i].category, resources[i].resourceName),
+      );
 
-    if (fs.pathExistsSync(targetDir)) {
-      fs.removeSync(targetDir);
+      if (fs.pathExistsSync(targetDir)) {
+        fs.removeSync(targetDir);
+      }
+
+      fs.ensureDirSync(targetDir);
+
+      fs.copySync(sourceDir, targetDir);
     }
-
-    fs.ensureDirSync(targetDir);
-
-    fs.copySync(sourceDir, targetDir);
   }
 
   fs.copySync(amplifyMetaFilePath, amplifyCloudMetaFilePath, { overwrite: true });
   fs.copySync(backendConfigFilePath, backendConfigCloudFilePath, { overwrite: true });
+
   // if project hasn't been initialized after tags has been released
   if (fs.existsSync(tagFilePath)) {
     fs.copySync(tagFilePath, tagCloudFilePath, { overwrite: true });
   }
 }
 
-export function updateamplifyMetaAfterResourceAdd(category, resourceName, options: { dependsOn? } = {}) {
+export function updateamplifyMetaAfterResourceAdd(
+  category,
+  resourceName,
+  metadataResource: { dependsOn? } = {},
+  backendConfigResource?: { dependsOn? },
+) {
   const amplifyMeta = stateManager.getMeta();
 
-  if (options.dependsOn) {
-    checkForCyclicDependencies(category, resourceName, options.dependsOn);
+  if (metadataResource.dependsOn) {
+    checkForCyclicDependencies(category, resourceName, metadataResource.dependsOn);
   }
 
   if (!amplifyMeta[category]) {
@@ -78,11 +86,14 @@ export function updateamplifyMetaAfterResourceAdd(category, resourceName, option
     throw new Error(`${resourceName} is present in amplify-meta.json`);
   }
   amplifyMeta[category][resourceName] = {};
-  amplifyMeta[category][resourceName] = options;
+  amplifyMeta[category][resourceName] = metadataResource;
 
   stateManager.setMeta(undefined, amplifyMeta);
 
-  updateBackendConfigAfterResourceAdd(category, resourceName, options);
+  // If a backend config resource passed in store it, otherwise the same data as in meta
+  // In case of imported resources the output block contains only the user selected values that
+  // are needed for recreation of sensitive data like secrets and such.
+  updateBackendConfigAfterResourceAdd(category, resourceName, backendConfigResource || metadataResource);
 }
 
 export function updateProvideramplifyMeta(providerName, options) {
@@ -124,13 +135,24 @@ export async function updateamplifyMetaAfterPush(resources) {
   const currentTimestamp = new Date();
 
   for (let i = 0; i < resources.length; i += 1) {
-    const sourceDir = path.normalize(path.join(pathManager.getBackendDirPath(), resources[i].category, resources[i].resourceName));
-    const hashDir = await getHashForResourceDir(sourceDir);
+    // Skip hash calculation for imported resources
+    if (resources[i].serviceType !== 'imported') {
+      const sourceDir = path.normalize(path.join(pathManager.getBackendDirPath(), resources[i].category, resources[i].resourceName));
+      const hashDir = await getHashForResourceDir(sourceDir);
 
-    /*eslint-disable */
-    amplifyMeta[resources[i].category][resources[i].resourceName].lastPushTimeStamp = currentTimestamp;
-    amplifyMeta[resources[i].category][resources[i].resourceName].lastPushDirHash = hashDir;
-    /* eslint-enable */
+      amplifyMeta[resources[i].category][resources[i].resourceName].lastPushDirHash = hashDir;
+      amplifyMeta[resources[i].category][resources[i].resourceName].lastPushTimeStamp = currentTimestamp;
+    }
+
+    // If the operation was a remove-sync then for imported resources we cannot set timestamp
+    // but those are still in the received array as this method is operation agnostic.
+    if (
+      resources[i].serviceType === 'imported' &&
+      amplifyMeta[resources[i].category] &&
+      amplifyMeta[resources[i].category][resources[i].resourceName]
+    ) {
+      amplifyMeta[resources[i].category][resources[i].resourceName].lastPushTimeStamp = currentTimestamp;
+    }
   }
 
   stateManager.setMeta(undefined, amplifyMeta);
